pyproject.toml
[tool.poetry]
name = "diaweb-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false # use it only to manage dependencies, not to create a package

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
absl-py = "2.1.0"
annotated-types = "0.7.0"
anyio = "4.4.0"
asgiref = "3.7.2"
astroid = "3.2.2"
astunparse = "1.6.3"
autopep8 = "2.2.0"
certifi = "2024.6.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cryptography = "42.0.8"
deprecation = "2.1.0"
dill = "0.3.8"
django = "5.0"
django-cors-headers = "4.3.1"
django-environ = "0.11.2"
djangorestframework = "3.15.1"
drf-yasg = "1.21.7"
et-xmlfile = "1.1.0"
flatbuffers = "24.3.25"
gast = "0.5.4"
gotrue = "2.4.4"
grpcio = "1.64.1"
grpcio-status = "1.62.2"
h11 = "0.14.0"
h5py = "3.11.0"
httpcore = "1.0.5"
httplib2 = "0.22.0"
httpx = "0.27.0"
idna = "3.7"
inflection = "0.5.1"
isort = "5.13.2"
keras = "3.3.3"
keras-tuner = "1.4.7"
kt-legacy = "1.0.5"
libclang = "18.1.1"
markdown = "3.6"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mccabe = "0.7.0"
mdurl = "0.1.2"
ml-dtypes = "0.3.2"
msgpack = "1.0.8"
namex = "0.0.8"
numpy = "1.26.4"
openpyxl = "3.1.4"
opt-einsum = "3.3.0"
optree = "0.11.0"
packaging = "24.0"
pandas = "2.2.2"
platformdirs = "4.2.2"
postgrest = "0.16.8"
proto-plus = "1.23.0"
protobuf = "4.25.3"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
pycodestyle = "2.11.1"
pycparser = "2.22"
pydantic = "2.7.3"
pydantic-core = "2.18.4"
pygments = "2.18.0"
pyjwt = "2.8.0"
pyparsing = "3.1.2"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.1"
pyyaml = "6.0.1"
realtime = "1.0.5"
requests = "2.32.3"
rich = "13.7.1"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.1"
sqlparse = "0.4.4"
storage3 = "0.7.6"
strenum = "0.4.15"
supafunc = "0.4.6"
tensorboard = "2.16.2"
tensorboard-data-server = "0.7.2"
tensorflow = "2.16.1"
tensorflow-io-gcs-filesystem = "0.31.0"
termcolor = "2.4.0"
tomlkit = "0.12.5"
typing-extensions = "4.12.2"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
websockets = "12.0"
werkzeug = "3.0.3"
wrapt = "1.16.0"
tensorflow-intel = { version = "^2.16.2", platform = "windows" } # Conditional dependency for Windows
channels = "^4.1.0"
daphne = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
